Rework atomic wrappers
- replace core::Atomic with core::Atomic<>
- use full barriers by default
- add explicit methods
- introduce AtomicOps layer, implemented using C11 and libatomic_ops
- libatomic_ops implementation is now generated
- by default, use C11 implementation if supported

+++ src/modules/roc_netio/target_libuv/roc_netio/network_loop.cpp
@@ -453,7 +453,7 @@ void NetworkLoop::finish_task_(Task& task) {
 }
 
 void NetworkLoop::update_num_ports_() {
-    num_open_ports_ = (long)open_ports_.size();
+    num_open_ports_ = (int)open_ports_.size();
 }
 
 } // namespace netio
